Updated Jade to Pug. Removed /pt from url locations.
The purpose of /pt was to differentiate between node instances on a
server. This will be replaced with an nginx configuration that proxies
requests for /pt to the node server. It will also add /pt to all links
that go through the proxy so the browser can correctly get the needed
resources.